backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
biopython==1.79
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.47.2
idna==3.6
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.23.5
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
ProDy==2.4.1
pyfastx==2.0.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
zipp==3.17.0
